// src/modules/reporting/dto/report.money.in.pipe.raw.dto.ts
export class ReportMoneyInPipeRawDto {

        offeredSalary: number;

        offeredCurrency: number;

        vacancyReference: string;

        offerStatus: string;

        offerLastUpdated: string;

        offerCreateDate: string;

        jobTitle: string;

        contractType: string;

        jobType: string;

        locationId: number;

        maxSalary: number;

        minSalary: number;

        currencyType: string;

        vacancyCreateDate: string;

        vacancyUpdatedDate: string;

        hourlyRate: number;

        weeklyRate: number;

        monthlyRate: number;

        jobDescription: JSON;

        commision: number

}